Disallow putting split results in nonexistent parent folder
The join method doesn't create the directory to put its result in,
so for consistency neither should the split function, except insofar
as the result is a directory itself.

This means that in the doctest, ".test" must already exist before split
is called, just as it must before join is.